backend_django/requirements.txt
# Django core
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0
dj-database-url>=2.0.0

# Database
psycopg2-binary>=2.9.0
django-extensions>=3.2.0

# Authentication & Security
djangorestframework-simplejwt>=5.2.0
django-allauth>=0.54.0
cryptography>=41.0.0

# Celery & Redis
celery>=5.3.0
redis>=4.5.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# API Documentation
drf-spectacular>=0.26.0
drf-yasg>=1.21.0

# Image & File handling
Pillow>=10.0.0
django-storages>=1.13.0
boto3>=1.28.0

# OpenAI & ML
openai>=1.6.1
anthropic>=0.7.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# WebSockets
channels>=4.0.0
channels-redis>=4.1.0

# Utilities
python-decouple>=3.8
django-environ>=0.10.0
python-dotenv>=1.0.0
requests>=2.31.0

# Development
django-debug-toolbar>=4.1.0
pytest-django>=4.5.0
factory-boy>=3.3.0
freezegun>=1.2.0

# Production
gunicorn>=20.1.0
whitenoise>=6.5.0
sentry-sdk>=1.28.0

# Math & Statistics (for IRT analysis)
scipy>=1.11.0
matplotlib>=3.7.0 